refactor: remove unused functions
