+++ test/helpers.test.ts
@@ -1,6 +1,6 @@
 import path from 'path'
 
-import { getGlobals, getRangeByLoc, requirePkg } from 'eslint-plugin-mdx'
+import { getGlobals, requirePkg } from 'eslint-plugin-mdx'
 
 describe('Helpers', () => {
   it('should resolve globals correctly', () => {
@@ -22,28 +22,4 @@ describe('Helpers', () => {
 
   it('should throw on non existed package', () =>
     expect(() => requirePkg('@1stg/config', 'unexpected-')).toThrow())
-})
-
-describe('getRangeByLoc', () => {
-  it('should work as expected', () => {
-    expect(
-      getRangeByLoc(
-        `
-export const a = 1
-export const a = 1
-      `.trim(),
-        {
-          start: {
-            line: 2,
-            column: 8,
-          },
-          end: {
-            line: 2,
-            column: 15,
-          },
-        },
-      ),
-      // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-    ).toEqual([26, 33])
-  })
 })
