+++ tsconfig.base.json
@@ -1,7 +1,6 @@
 {
   "extends": "./node_modules/@1stg/tsconfig/lib.json",
   "compilerOptions": {
-    "baseUrl": ".",
     "declarationMap": false,
     "strictFunctionTypes": false,
     "strictNullChecks": false
